+++ background.worker.js
@@ -162,7 +162,7 @@ function addListeners() {
   toggleListener(
     browser.tabs.onUpdated,
     onTabUpdated,
-    hasAny(['tabs.pinned', 'tabs.unpinned'], types)
+    hasAny(['tabs.pinned', 'tabs.unpinned', 'tabs.group-in', 'tabs.group-out'], types)
     // {
     //   urls: ['<all_urls>'],
     //   properties: ['attention', 'pinned']
@@ -341,6 +341,14 @@ function onTabUpdated(tabId, changeInfo, tabInfo) {
       play('tabs.pinned');
     } else {
       play('tabs.unpinned');
+    }
+  }
+
+  if (keys.includes('groupId')) {
+    if (changeInfo['groupId'] === -1) {
+      play('tabs.group-out');
+    } else {
+      play('tabs.group-in');
     }
   }
 }