+++ CHANGES.md
@@ -4,6 +4,7 @@ CHANGES
 ## 3.?.? (?)
 
 * Add tab groups open / close / move / update events.
+* Add tab add-into / remove-from group events.
 * Fix missing to remove tabs.onTabUpdated listener on Firefox.
 
 ## 3.3.0 (2025-06-23)